+++ src/apps/Loop/Sections/AnalyticsConfig.tsx
@@ -1,6 +1,6 @@
 import { GardenOptions } from '../../../components/CompletionView/src/DataViewerApi/DataViewState';
 import { Garden } from '../../../components/ParkView/Components/Garden';
-import { AnalyticsOptions } from '../../../packages/Diagrams/src/types/analyticsOptions';
+import { AnalyticsOptions } from '../../../packages/Diagrams/src/2types/analyticsOptions';
 import {
     getDatePercentage,
     getDateStatus,
