+++ src/packages/Diagrams/src/Components/GetChartsStyles.ts
