+++ src/packages/Diagrams/src/Types/cumulativeSeries.ts
