+++ src/packages/Diagrams/src/Types/cumulativeSeriesOptions.ts
