+++ src/packages/Diagrams/src/Utils/sortByKey.ts
@@ -8,7 +8,7 @@
  * @param {boolean} [accenting=true]
  * @return {*}
  */
-export function sortByKey<T>(dataItem: T[], key: keyof T, accenting: boolean = true) {
+export function sortByKey<T>(dataItem: T[], key: keyof T, accenting = true): T[] {
     return dataItem.sort((a: T, b: T) => {
         if (accenting) {
             return a[key as string] - b[key as string];
