+++ src/packages/Diagrams/src/Visuals/BarVisual/BarVisual.tsx
