+++ src/packages/Diagrams/src/Visuals/BarVisual/Hooks/useBarChart.ts
