+++ src/packages/Diagrams/src/Visuals/BarVisual/Readme.md
