+++ src/packages/Diagrams/src/Visuals/BarVisual/Types/barVisualOptions.ts
