+++ src/packages/Diagrams/src/Visuals/BarVisual/index.ts
