+++ src/packages/Diagrams/src/Visuals/LineVisual/index.ts
