+++ src/packages/Diagrams/src/Visuals/TableVisual/Styles/Table.ts
