+++ src/packages/Diagrams/src/Visuals/TableVisual/Types/table.ts
