+++ src/packages/Diagrams/src/Visuals/TableVisual/Types/tableVisualOptions.ts
