+++ src/packages/Diagrams/src/Visuals/TableVisual/Utils/generateDefaultColumns.tsx
