+++ src/packages/Diagrams/src/Visuals/TimeVisual/Styles/Styles.ts
