+++ src/packages/Diagrams/src/index.ts
@@ -1,3 +1,3 @@
-export * from './2Components/AnalyticsView';
-export * from './2types/analyticsOptions';
-export * from './2Visuals';
+export * from './Components/AnalyticsView';
+export * from './Types';
+export * from './Visuals';
